refactor(mongo): relocate Mongo TLS PEM lifecycle to neutral internal/adapters/mongo_tls (spec 042 / PRD 29)
Phase D of the ADR 0001 cleanup — removes the LAST adapter→adapter cross-axis
import. The Mongo TLS PEM lifecycle (MongoTLSMaterial, PrepareMongoTLS,
Register/Unregister/CloseAll, SweepOrphanMaterial) lived in the dump engine
adapter internal/adapters/dump/mongo/ but was consumed by the restore Mongo
adapter and CLI startup/shutdown — the only restore/mongo → dump/mongo import.

- Move material.go, cleanup.go, sweep.go (+ their tests) to the neutral package
  internal/adapters/mongo_tls (package mongo_tls) + add doc.go. Package-level
  functions over the singleton registry — no port glue (clarify Option A).
- Swap imports: cli/root.go and restore/mongo/args_builder.go (alias target only,
  call sites unchanged — restore/mongo no longer imports dump/mongo);
  dump/mongo/args_builder.go now imports mongo_tls and qualifies
  PrepareMongoTLS/MongoTLSMaterial.
- .golangci.yml: reword the adapter-restore-axis comment (no allowlist entry
  existed). CLAUDE.md: three refs updated; the cross-axis Mongo PEM exception
  bullet removed — only dump/* → db_probe (PRD 30) remains.

Pure relocation, behaviour byte-identical: same PEM files, args, single shared
registry, sweep. build/vet/test green; make lint 0; make e2e 32 PASS / 0 FAIL
(incl. mongo path). Refs denisakp/sentinel#69.
