+++ internal/adapters/mongo_tls/material.go
@@ -1,4 +1,4 @@
-package mongo
+package mongo_tls
 
 import (
 	"bytes"
