+++ internal/adapters/mongo_tls/material_test.go
@@ -1,4 +1,4 @@
-package mongo
+package mongo_tls
 
 import (
 	"bytes"
