+++ internal/adapters/mongo_tls/sweep.go
@@ -1,4 +1,4 @@
-package mongo
+package mongo_tls
 
 import (
 	"errors"
